feat(build): runtime-free primitive registrar generator (lgprimgen)
The registrar in pkg/rt/zz_primitives_generated.go is what lets the runtime
boot; a generator that boots the runtime to produce it depends on the very
artifact it emits (p0 depending on p1). When a primitive is hoisted out of
installLangNS into a //lg:native decl, the runtime cannot boot until this file
carries its registration — so the generator must run in exactly that state.

- Extract the primitives codegen into internal/primgen, a PURE go/ast tool
  with no pkg/compiler / pkg/rt import (prims_scan, prims_emit, Generate,
  ScanNativeNames, KebabCase).
- Add cmd/lgprimgen, a thin binary over primgen; cmd/lginterop delegates its
  -primitives mode to the same package (its compiler import is only for the
  interop-EDN path).
- Makefile: regenerate zz_primitives_generated.go via lgprimgen BEFORE the
  lgbgen bootstrap that consumes it, breaking the boot cycle. generate.lg
  calls lgprimgen too.
- Fix a latent emit bug: variadic-spread adapters (Fn(vs...)) dropped the
  closing paren — never exercised until the first //lg:native ...vm.Value
  primitives. gofmt failures now surface stderr + dump the source.
