+++ .gitignore
@@ -175,6 +175,7 @@ AGENTS.md
 # AOT codegen tool — built ad-hoc, not checked in
 /lgbgen
 /lginterop
+/lgprimgen
 /bench-ratchet
 /perf-page
