+++ scripts/generate.lg
@@ -141,9 +141,15 @@
 (gen-go   "scripts/gen/ir_bridge_generated.head" "pkg/ir/ir_bridge.lg" "pkg/rt/ir_bridge_generated.go")
 (gen-lisp "pkg/ir/ir_data.lg" "pkg/rt/core/ir/data/generated.lg")
 
-;; Invoke cmd/lginterop -primitives to generate the primitive registrar.
-;; Scans pkg/rt for //lg:native-annotated functions and generates adapters + registrations.
-(run! go-bin "run" "./cmd/lginterop" "-primitives" "pkg/rt" "-go-pkg" "github.com/nooga/let-go/pkg/rt" "-primitives-out" "pkg/rt/zz_primitives_generated.go")
+;; Invoke cmd/lgprimgen to generate the primitive registrar.
+;; Scans pkg/rt for //lg:native-annotated functions and generates adapters +
+;; registrations. lgprimgen is a PURE go/ast codegen tool that does NOT import
+;; the runtime (pkg/compiler / pkg/rt) — so it runs even mid-migration, when
+;; primitives have been moved out of installLangNS but the registrar that
+;; replaces them has not been generated yet and the runtime cannot boot. Using
+;; the runtime-coupled cmd/lginterop here would make this step depend on the
+;; very artifact it produces.
+(run! go-bin "run" "./cmd/lgprimgen" "-primitives" "pkg/rt" "-go-pkg" "github.com/nooga/let-go/pkg/rt" "-primitives-out" "pkg/rt/zz_primitives_generated.go")
 
 ;; One core compile emits BOTH the .lgb bundle and the gogen_ir Go tree.
 ;; Hermetic generation (clearing the stale lowered-tree before regen) is
